@@ -1,11 +1,9 @@
 module github.com/Azure/azure-sdk-for-go/sdk/compute/armcompute
 
-go 1.13
+go 1.16
 
 require (
-	github.com/Azure/azure-sdk-for-go v55.2.0+incompatible
-	github.com/Azure/azure-sdk-for-go/sdk/armcore v0.8.0
-	github.com/Azure/azure-sdk-for-go/sdk/azcore v0.16.2
-	github.com/Azure/azure-sdk-for-go/sdk/azidentity v0.9.1
-	github.com/Azure/azure-sdk-for-go/sdk/to v0.1.4
+	github.com/Azure/azure-sdk-for-go v57.0.0+incompatible
+	github.com/Azure/azure-sdk-for-go/sdk/azcore v0.19.0
+	github.com/Azure/azure-sdk-for-go/sdk/azidentity v0.10.0
 )
