@@ -1,4 +1,5 @@
-// +build go1.13
+//go:build go1.16
+// +build go1.16
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -9,10 +10,10 @@ import (
 	"context"
 	"log"
 
-	"github.com/Azure/azure-sdk-for-go/sdk/armcore"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
 	"github.com/Azure/azure-sdk-for-go/sdk/compute/armcompute"
-	"github.com/Azure/azure-sdk-for-go/sdk/to"
 )
 
 func ExampleAvailabilitySetsClient_CreateOrUpdate() {
@@ -20,7 +21,7 @@ func ExampleAvailabilitySetsClient_CreateOrUpdate() {
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
-	client := armcompute.NewAvailabilitySetsClient(armcore.NewDefaultConnection(cred, nil), "<subscription ID>")
+	client := armcompute.NewAvailabilitySetsClient(arm.NewDefaultConnection(cred, nil), "<subscription ID>")
 	resp, err := client.CreateOrUpdate(
 		context.Background(),
 		"<resource group name>",
@@ -51,7 +52,7 @@ func ExampleAvailabilitySetsClient_Get() {
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
-	client := armcompute.NewAvailabilitySetsClient(armcore.NewDefaultConnection(cred, nil), "<subscription ID>")
+	client := armcompute.NewAvailabilitySetsClient(arm.NewDefaultConnection(cred, nil), "<subscription ID>")
 	resp, err := client.Get(context.Background(), "<resource group name>", "<availability set name>", nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a response: %v", err)
