@@ -1,4 +1,5 @@
-// +build go1.13
+//go:build go1.16
+// +build go1.16
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -10,10 +11,10 @@ import (
 	"log"
 	"time"
 
-	"github.com/Azure/azure-sdk-for-go/sdk/armcore"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
 	"github.com/Azure/azure-sdk-for-go/sdk/compute/armcompute"
-	"github.com/Azure/azure-sdk-for-go/sdk/to"
 )
 
 func ExampleDisksClient_BeginCreateOrUpdate() {
@@ -21,7 +22,7 @@ func ExampleDisksClient_BeginCreateOrUpdate() {
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
-	client := armcompute.NewDisksClient(armcore.NewDefaultConnection(cred, nil), "<subscription ID>")
+	client := armcompute.NewDisksClient(arm.NewDefaultConnection(cred, nil), "<subscription ID>")
 	poller, err := client.BeginCreateOrUpdate(
 		context.Background(),
 		"<resource group name>",
@@ -55,7 +56,7 @@ func ExampleDisksClient_Get() {
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
-	client := armcompute.NewDisksClient(armcore.NewDefaultConnection(cred, nil), "<subscription ID>")
+	client := armcompute.NewDisksClient(arm.NewDefaultConnection(cred, nil), "<subscription ID>")
 	resp, err := client.Get(context.Background(), "<resource group name>", "<disk name>", nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a response: %v", err)
