@@ -1,4 +1,5 @@
-// +build go1.13
+//go:build go1.16
+// +build go1.16
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -11,10 +12,10 @@ import (
 	"log"
 	"time"
 
-	"github.com/Azure/azure-sdk-for-go/sdk/armcore"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
 	"github.com/Azure/azure-sdk-for-go/sdk/compute/armcompute"
-	"github.com/Azure/azure-sdk-for-go/sdk/to"
 )
 
 func ExampleVirtualMachineExtensionsClient_BeginCreateOrUpdate() {
@@ -22,7 +23,7 @@ func ExampleVirtualMachineExtensionsClient_BeginCreateOrUpdate() {
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
-	client := armcompute.NewVirtualMachineExtensionsClient(armcore.NewDefaultConnection(cred, nil), "<subscription ID>")
+	client := armcompute.NewVirtualMachineExtensionsClient(arm.NewDefaultConnection(cred, nil), "<subscription ID>")
 	poller, err := client.BeginCreateOrUpdate(
 		context.Background(),
 		"<resource group name>",
@@ -69,7 +70,7 @@ func ExampleVirtualMachineExtensionsClient_Get() {
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
-	client := armcompute.NewVirtualMachineExtensionsClient(armcore.NewDefaultConnection(cred, nil), "<subscription ID>")
+	client := armcompute.NewVirtualMachineExtensionsClient(arm.NewDefaultConnection(cred, nil), "<subscription ID>")
 	resp, err := client.Get(context.Background(), "<resource group name>", "<VM name>", "<VM extension name>", nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a response: %v", err)
