@@ -1,4 +1,5 @@
-// +build go1.13
+//go:build go1.16
+// +build go1.16
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -11,10 +12,10 @@ import (
 	"log"
 	"time"
 
-	"github.com/Azure/azure-sdk-for-go/sdk/armcore"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
 	"github.com/Azure/azure-sdk-for-go/sdk/compute/armcompute"
-	"github.com/Azure/azure-sdk-for-go/sdk/to"
 )
 
 func ExampleVirtualMachinesClient_BeginCreateOrUpdate() {
@@ -24,7 +25,7 @@ func ExampleVirtualMachinesClient_BeginCreateOrUpdate() {
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
-	client := armcompute.NewVirtualMachinesClient(armcore.NewDefaultConnection(cred, nil), "<subscription ID>")
+	client := armcompute.NewVirtualMachinesClient(arm.NewDefaultConnection(cred, nil), "<subscription ID>")
 	poller, err := client.BeginCreateOrUpdate(
 		context.Background(),
 		"<resource group name>",
@@ -94,7 +95,7 @@ func ExampleVirtualMachinesClient_BeginCreateOrUpdate_withDisk() {
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
-	client := armcompute.NewVirtualMachinesClient(armcore.NewDefaultConnection(cred, nil), "<subscription ID>")
+	client := armcompute.NewVirtualMachinesClient(arm.NewDefaultConnection(cred, nil), "<subscription ID>")
 	poller, err := client.BeginCreateOrUpdate(
 		context.Background(),
 		"<resource group name>",
@@ -172,7 +173,7 @@ func ExampleVirtualMachinesClient_BeginCreateOrUpdate_withLoadBalancer() {
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
-	client := armcompute.NewVirtualMachinesClient(armcore.NewDefaultConnection(cred, nil), "<subscription ID>")
+	client := armcompute.NewVirtualMachinesClient(arm.NewDefaultConnection(cred, nil), "<subscription ID>")
 	poller, err := client.BeginCreateOrUpdate(
 		context.Background(),
 		"<resource group name>",
@@ -235,7 +236,7 @@ func ExampleVirtualMachinesClient_Get() {
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
-	client := armcompute.NewVirtualMachinesClient(armcore.NewDefaultConnection(cred, nil), "<subscription ID>")
+	client := armcompute.NewVirtualMachinesClient(arm.NewDefaultConnection(cred, nil), "<subscription ID>")
 	resp, err := client.Get(context.Background(), "<resource group name>", "<VM name>", nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a response: %v", err)
@@ -248,7 +249,7 @@ func ExampleVirtualMachinesClient_BeginDeallocate() {
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
-	client := armcompute.NewVirtualMachinesClient(armcore.NewDefaultConnection(cred, nil), "<subscription ID>")
+	client := armcompute.NewVirtualMachinesClient(arm.NewDefaultConnection(cred, nil), "<subscription ID>")
 	poller, err := client.BeginDeallocate(context.Background(), "<resource group name>", "<VM name>", nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a response: %v", err)
@@ -264,7 +265,7 @@ func ExampleVirtualMachinesClient_BeginUpdate() {
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
-	client := armcompute.NewVirtualMachinesClient(armcore.NewDefaultConnection(cred, nil), "<subscription ID>")
+	client := armcompute.NewVirtualMachinesClient(arm.NewDefaultConnection(cred, nil), "<subscription ID>")
 	poller, err := client.BeginUpdate(
 		context.Background(),
 		"<resource group name>",
@@ -294,7 +295,7 @@ func ExampleVirtualMachinesClient_BeginStart() {
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
-	client := armcompute.NewVirtualMachinesClient(armcore.NewDefaultConnection(cred, nil), "<subscription ID>")
+	client := armcompute.NewVirtualMachinesClient(arm.NewDefaultConnection(cred, nil), "<subscription ID>")
 	poller, err := client.BeginStart(
 		context.Background(),
 		"<resource group name>",
@@ -315,7 +316,7 @@ func ExampleVirtualMachinesClient_BeginRestart() {
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
-	client := armcompute.NewVirtualMachinesClient(armcore.NewDefaultConnection(cred, nil), "<subscription ID>")
+	client := armcompute.NewVirtualMachinesClient(arm.NewDefaultConnection(cred, nil), "<subscription ID>")
 	poller, err := client.BeginRestart(
 		context.Background(),
 		"<resource group name>",
@@ -336,7 +337,7 @@ func ExampleVirtualMachinesClient_BeginPowerOff() {
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
-	client := armcompute.NewVirtualMachinesClient(armcore.NewDefaultConnection(cred, nil), "<subscription ID>")
+	client := armcompute.NewVirtualMachinesClient(arm.NewDefaultConnection(cred, nil), "<subscription ID>")
 	poller, err := client.BeginPowerOff(
 		context.Background(),
 		"<resource group name>",
