@@ -1,4 +1,5 @@
-// +build go1.13
+//go:build go1.16
+// +build go1.16
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -11,44 +12,47 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"github.com/Azure/azure-sdk-for-go/sdk/armcore"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"net/http"
 	"net/url"
 	"strings"
+
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 )
 
 // CloudServiceOperatingSystemsClient contains the methods for the CloudServiceOperatingSystems group.
 // Don't use this type directly, use NewCloudServiceOperatingSystemsClient() instead.
 type CloudServiceOperatingSystemsClient struct {
-	con            *armcore.Connection
+	ep             string
+	pl             runtime.Pipeline
 	subscriptionID string
 }
 
 // NewCloudServiceOperatingSystemsClient creates a new instance of CloudServiceOperatingSystemsClient with the specified values.
-func NewCloudServiceOperatingSystemsClient(con *armcore.Connection, subscriptionID string) *CloudServiceOperatingSystemsClient {
-	return &CloudServiceOperatingSystemsClient{con: con, subscriptionID: subscriptionID}
+func NewCloudServiceOperatingSystemsClient(con *arm.Connection, subscriptionID string) *CloudServiceOperatingSystemsClient {
+	return &CloudServiceOperatingSystemsClient{ep: con.Endpoint(), pl: con.NewPipeline(module, version), subscriptionID: subscriptionID}
 }
 
 // GetOSFamily - Gets properties of a guest operating system family that can be specified in the XML service configuration (.cscfg) for a cloud service.
 // If the operation fails it returns the *CloudError error type.
-func (client *CloudServiceOperatingSystemsClient) GetOSFamily(ctx context.Context, location string, osFamilyName string, options *CloudServiceOperatingSystemsGetOSFamilyOptions) (OSFamilyResponse, error) {
+func (client *CloudServiceOperatingSystemsClient) GetOSFamily(ctx context.Context, location string, osFamilyName string, options *CloudServiceOperatingSystemsGetOSFamilyOptions) (CloudServiceOperatingSystemsGetOSFamilyResponse, error) {
 	req, err := client.getOSFamilyCreateRequest(ctx, location, osFamilyName, options)
 	if err != nil {
-		return OSFamilyResponse{}, err
+		return CloudServiceOperatingSystemsGetOSFamilyResponse{}, err
 	}
-	resp, err := client.con.Pipeline().Do(req)
+	resp, err := client.pl.Do(req)
 	if err != nil {
-		return OSFamilyResponse{}, err
+		return CloudServiceOperatingSystemsGetOSFamilyResponse{}, err
 	}
-	if !resp.HasStatusCode(http.StatusOK) {
-		return OSFamilyResponse{}, client.getOSFamilyHandleError(resp)
+	if !runtime.HasStatusCode(resp, http.StatusOK) {
+		return CloudServiceOperatingSystemsGetOSFamilyResponse{}, client.getOSFamilyHandleError(resp)
 	}
 	return client.getOSFamilyHandleResponse(resp)
 }
 
 // getOSFamilyCreateRequest creates the GetOSFamily request.
-func (client *CloudServiceOperatingSystemsClient) getOSFamilyCreateRequest(ctx context.Context, location string, osFamilyName string, options *CloudServiceOperatingSystemsGetOSFamilyOptions) (*azcore.Request, error) {
+func (client *CloudServiceOperatingSystemsClient) getOSFamilyCreateRequest(ctx context.Context, location string, osFamilyName string, options *CloudServiceOperatingSystemsGetOSFamilyOptions) (*policy.Request, error) {
 	urlPath := "/subscriptions/{subscriptionId}/providers/Microsoft.Compute/locations/{location}/cloudServiceOsFamilies/{osFamilyName}"
 	if location == "" {
 		return nil, errors.New("parameter location cannot be empty")
@@ -62,59 +66,58 @@ func (client *CloudServiceOperatingSystemsClient) getOSFamilyCreateRequest(ctx c
 		return nil, errors.New("parameter client.subscriptionID cannot be empty")
 	}
 	urlPath = strings.ReplaceAll(urlPath, "{subscriptionId}", url.PathEscape(client.subscriptionID))
-	req, err := azcore.NewRequest(ctx, http.MethodGet, azcore.JoinPaths(client.con.Endpoint(), urlPath))
+	req, err := runtime.NewRequest(ctx, http.MethodGet, runtime.JoinPaths(client.ep, urlPath))
 	if err != nil {
 		return nil, err
 	}
-	req.Telemetry(telemetryInfo)
-	reqQP := req.URL.Query()
+	reqQP := req.Raw().URL.Query()
 	reqQP.Set("api-version", "2021-03-01")
-	req.URL.RawQuery = reqQP.Encode()
-	req.Header.Set("Accept", "application/json")
+	req.Raw().URL.RawQuery = reqQP.Encode()
+	req.Raw().Header.Set("Accept", "application/json")
 	return req, nil
 }
 
 // getOSFamilyHandleResponse handles the GetOSFamily response.
-func (client *CloudServiceOperatingSystemsClient) getOSFamilyHandleResponse(resp *azcore.Response) (OSFamilyResponse, error) {
-	var val *OSFamily
-	if err := resp.UnmarshalAsJSON(&val); err != nil {
-		return OSFamilyResponse{}, err
+func (client *CloudServiceOperatingSystemsClient) getOSFamilyHandleResponse(resp *http.Response) (CloudServiceOperatingSystemsGetOSFamilyResponse, error) {
+	result := CloudServiceOperatingSystemsGetOSFamilyResponse{RawResponse: resp}
+	if err := runtime.UnmarshalAsJSON(resp, &result.OSFamily); err != nil {
+		return CloudServiceOperatingSystemsGetOSFamilyResponse{}, err
 	}
-	return OSFamilyResponse{RawResponse: resp.Response, OSFamily: val}, nil
+	return result, nil
 }
 
 // getOSFamilyHandleError handles the GetOSFamily error response.
-func (client *CloudServiceOperatingSystemsClient) getOSFamilyHandleError(resp *azcore.Response) error {
-	body, err := resp.Payload()
+func (client *CloudServiceOperatingSystemsClient) getOSFamilyHandleError(resp *http.Response) error {
+	body, err := runtime.Payload(resp)
 	if err != nil {
-		return azcore.NewResponseError(err, resp.Response)
+		return runtime.NewResponseError(err, resp)
 	}
 	errType := CloudError{raw: string(body)}
-	if err := resp.UnmarshalAsJSON(&errType); err != nil {
-		return azcore.NewResponseError(fmt.Errorf("%s\n%s", string(body), err), resp.Response)
+	if err := runtime.UnmarshalAsJSON(resp, &errType); err != nil {
+		return runtime.NewResponseError(fmt.Errorf("%s\n%s", string(body), err), resp)
 	}
-	return azcore.NewResponseError(&errType, resp.Response)
+	return runtime.NewResponseError(&errType, resp)
 }
 
 // GetOSVersion - Gets properties of a guest operating system version that can be specified in the XML service configuration (.cscfg) for a cloud service.
 // If the operation fails it returns the *CloudError error type.
-func (client *CloudServiceOperatingSystemsClient) GetOSVersion(ctx context.Context, location string, osVersionName string, options *CloudServiceOperatingSystemsGetOSVersionOptions) (OSVersionResponse, error) {
+func (client *CloudServiceOperatingSystemsClient) GetOSVersion(ctx context.Context, location string, osVersionName string, options *CloudServiceOperatingSystemsGetOSVersionOptions) (CloudServiceOperatingSystemsGetOSVersionResponse, error) {
 	req, err := client.getOSVersionCreateRequest(ctx, location, osVersionName, options)
 	if err != nil {
-		return OSVersionResponse{}, err
+		return CloudServiceOperatingSystemsGetOSVersionResponse{}, err
 	}
-	resp, err := client.con.Pipeline().Do(req)
+	resp, err := client.pl.Do(req)
 	if err != nil {
-		return OSVersionResponse{}, err
+		return CloudServiceOperatingSystemsGetOSVersionResponse{}, err
 	}
-	if !resp.HasStatusCode(http.StatusOK) {
-		return OSVersionResponse{}, client.getOSVersionHandleError(resp)
+	if !runtime.HasStatusCode(resp, http.StatusOK) {
+		return CloudServiceOperatingSystemsGetOSVersionResponse{}, client.getOSVersionHandleError(resp)
 	}
 	return client.getOSVersionHandleResponse(resp)
 }
 
 // getOSVersionCreateRequest creates the GetOSVersion request.
-func (client *CloudServiceOperatingSystemsClient) getOSVersionCreateRequest(ctx context.Context, location string, osVersionName string, options *CloudServiceOperatingSystemsGetOSVersionOptions) (*azcore.Request, error) {
+func (client *CloudServiceOperatingSystemsClient) getOSVersionCreateRequest(ctx context.Context, location string, osVersionName string, options *CloudServiceOperatingSystemsGetOSVersionOptions) (*policy.Request, error) {
 	urlPath := "/subscriptions/{subscriptionId}/providers/Microsoft.Compute/locations/{location}/cloudServiceOsVersions/{osVersionName}"
 	if location == "" {
 		return nil, errors.New("parameter location cannot be empty")
@@ -128,61 +131,57 @@ func (client *CloudServiceOperatingSystemsClient) getOSVersionCreateRequest(ctx 
 		return nil, errors.New("parameter client.subscriptionID cannot be empty")
 	}
 	urlPath = strings.ReplaceAll(urlPath, "{subscriptionId}", url.PathEscape(client.subscriptionID))
-	req, err := azcore.NewRequest(ctx, http.MethodGet, azcore.JoinPaths(client.con.Endpoint(), urlPath))
+	req, err := runtime.NewRequest(ctx, http.MethodGet, runtime.JoinPaths(client.ep, urlPath))
 	if err != nil {
 		return nil, err
 	}
-	req.Telemetry(telemetryInfo)
-	reqQP := req.URL.Query()
+	reqQP := req.Raw().URL.Query()
 	reqQP.Set("api-version", "2021-03-01")
-	req.URL.RawQuery = reqQP.Encode()
-	req.Header.Set("Accept", "application/json")
+	req.Raw().URL.RawQuery = reqQP.Encode()
+	req.Raw().Header.Set("Accept", "application/json")
 	return req, nil
 }
 
 // getOSVersionHandleResponse handles the GetOSVersion response.
-func (client *CloudServiceOperatingSystemsClient) getOSVersionHandleResponse(resp *azcore.Response) (OSVersionResponse, error) {
-	var val *OSVersion
-	if err := resp.UnmarshalAsJSON(&val); err != nil {
-		return OSVersionResponse{}, err
+func (client *CloudServiceOperatingSystemsClient) getOSVersionHandleResponse(resp *http.Response) (CloudServiceOperatingSystemsGetOSVersionResponse, error) {
+	result := CloudServiceOperatingSystemsGetOSVersionResponse{RawResponse: resp}
+	if err := runtime.UnmarshalAsJSON(resp, &result.OSVersion); err != nil {
+		return CloudServiceOperatingSystemsGetOSVersionResponse{}, err
 	}
-	return OSVersionResponse{RawResponse: resp.Response, OSVersion: val}, nil
+	return result, nil
 }
 
 // getOSVersionHandleError handles the GetOSVersion error response.
-func (client *CloudServiceOperatingSystemsClient) getOSVersionHandleError(resp *azcore.Response) error {
-	body, err := resp.Payload()
+func (client *CloudServiceOperatingSystemsClient) getOSVersionHandleError(resp *http.Response) error {
+	body, err := runtime.Payload(resp)
 	if err != nil {
-		return azcore.NewResponseError(err, resp.Response)
+		return runtime.NewResponseError(err, resp)
 	}
 	errType := CloudError{raw: string(body)}
-	if err := resp.UnmarshalAsJSON(&errType); err != nil {
-		return azcore.NewResponseError(fmt.Errorf("%s\n%s", string(body), err), resp.Response)
+	if err := runtime.UnmarshalAsJSON(resp, &errType); err != nil {
+		return runtime.NewResponseError(fmt.Errorf("%s\n%s", string(body), err), resp)
 	}
-	return azcore.NewResponseError(&errType, resp.Response)
+	return runtime.NewResponseError(&errType, resp)
 }
 
 // ListOSFamilies - Gets a list of all guest operating system families available to be specified in the XML service configuration (.cscfg) for a cloud service.
 // Use nextLink property in the response to get the next page
 // of OS Families. Do this till nextLink is null to fetch all the OS Families.
 // If the operation fails it returns the *CloudError error type.
-func (client *CloudServiceOperatingSystemsClient) ListOSFamilies(location string, options *CloudServiceOperatingSystemsListOSFamiliesOptions) OSFamilyListResultPager {
-	return &osFamilyListResultPager{
-		pipeline: client.con.Pipeline(),
-		requester: func(ctx context.Context) (*azcore.Request, error) {
+func (client *CloudServiceOperatingSystemsClient) ListOSFamilies(location string, options *CloudServiceOperatingSystemsListOSFamiliesOptions) *CloudServiceOperatingSystemsListOSFamiliesPager {
+	return &CloudServiceOperatingSystemsListOSFamiliesPager{
+		client: client,
+		requester: func(ctx context.Context) (*policy.Request, error) {
 			return client.listOSFamiliesCreateRequest(ctx, location, options)
 		},
-		responder: client.listOSFamiliesHandleResponse,
-		errorer:   client.listOSFamiliesHandleError,
-		advancer: func(ctx context.Context, resp OSFamilyListResultResponse) (*azcore.Request, error) {
-			return azcore.NewRequest(ctx, http.MethodGet, *resp.OSFamilyListResult.NextLink)
+		advancer: func(ctx context.Context, resp CloudServiceOperatingSystemsListOSFamiliesResponse) (*policy.Request, error) {
+			return runtime.NewRequest(ctx, http.MethodGet, *resp.OSFamilyListResult.NextLink)
 		},
-		statusCodes: []int{http.StatusOK},
 	}
 }
 
 // listOSFamiliesCreateRequest creates the ListOSFamilies request.
-func (client *CloudServiceOperatingSystemsClient) listOSFamiliesCreateRequest(ctx context.Context, location string, options *CloudServiceOperatingSystemsListOSFamiliesOptions) (*azcore.Request, error) {
+func (client *CloudServiceOperatingSystemsClient) listOSFamiliesCreateRequest(ctx context.Context, location string, options *CloudServiceOperatingSystemsListOSFamiliesOptions) (*policy.Request, error) {
 	urlPath := "/subscriptions/{subscriptionId}/providers/Microsoft.Compute/locations/{location}/cloudServiceOsFamilies"
 	if location == "" {
 		return nil, errors.New("parameter location cannot be empty")
@@ -192,61 +191,57 @@ func (client *CloudServiceOperatingSystemsClient) listOSFamiliesCreateRequest(ct
 		return nil, errors.New("parameter client.subscriptionID cannot be empty")
 	}
 	urlPath = strings.ReplaceAll(urlPath, "{subscriptionId}", url.PathEscape(client.subscriptionID))
-	req, err := azcore.NewRequest(ctx, http.MethodGet, azcore.JoinPaths(client.con.Endpoint(), urlPath))
+	req, err := runtime.NewRequest(ctx, http.MethodGet, runtime.JoinPaths(client.ep, urlPath))
 	if err != nil {
 		return nil, err
 	}
-	req.Telemetry(telemetryInfo)
-	reqQP := req.URL.Query()
+	reqQP := req.Raw().URL.Query()
 	reqQP.Set("api-version", "2021-03-01")
-	req.URL.RawQuery = reqQP.Encode()
-	req.Header.Set("Accept", "application/json")
+	req.Raw().URL.RawQuery = reqQP.Encode()
+	req.Raw().Header.Set("Accept", "application/json")
 	return req, nil
 }
 
 // listOSFamiliesHandleResponse handles the ListOSFamilies response.
-func (client *CloudServiceOperatingSystemsClient) listOSFamiliesHandleResponse(resp *azcore.Response) (OSFamilyListResultResponse, error) {
-	var val *OSFamilyListResult
-	if err := resp.UnmarshalAsJSON(&val); err != nil {
-		return OSFamilyListResultResponse{}, err
+func (client *CloudServiceOperatingSystemsClient) listOSFamiliesHandleResponse(resp *http.Response) (CloudServiceOperatingSystemsListOSFamiliesResponse, error) {
+	result := CloudServiceOperatingSystemsListOSFamiliesResponse{RawResponse: resp}
+	if err := runtime.UnmarshalAsJSON(resp, &result.OSFamilyListResult); err != nil {
+		return CloudServiceOperatingSystemsListOSFamiliesResponse{}, err
 	}
-	return OSFamilyListResultResponse{RawResponse: resp.Response, OSFamilyListResult: val}, nil
+	return result, nil
 }
 
 // listOSFamiliesHandleError handles the ListOSFamilies error response.
-func (client *CloudServiceOperatingSystemsClient) listOSFamiliesHandleError(resp *azcore.Response) error {
-	body, err := resp.Payload()
+func (client *CloudServiceOperatingSystemsClient) listOSFamiliesHandleError(resp *http.Response) error {
+	body, err := runtime.Payload(resp)
 	if err != nil {
-		return azcore.NewResponseError(err, resp.Response)
+		return runtime.NewResponseError(err, resp)
 	}
 	errType := CloudError{raw: string(body)}
-	if err := resp.UnmarshalAsJSON(&errType); err != nil {
-		return azcore.NewResponseError(fmt.Errorf("%s\n%s", string(body), err), resp.Response)
+	if err := runtime.UnmarshalAsJSON(resp, &errType); err != nil {
+		return runtime.NewResponseError(fmt.Errorf("%s\n%s", string(body), err), resp)
 	}
-	return azcore.NewResponseError(&errType, resp.Response)
+	return runtime.NewResponseError(&errType, resp)
 }
 
 // ListOSVersions - Gets a list of all guest operating system versions available to be specified in the XML service configuration (.cscfg) for a cloud service.
 // Use nextLink property in the response to get the next page
 // of OS versions. Do this till nextLink is null to fetch all the OS versions.
 // If the operation fails it returns the *CloudError error type.
-func (client *CloudServiceOperatingSystemsClient) ListOSVersions(location string, options *CloudServiceOperatingSystemsListOSVersionsOptions) OSVersionListResultPager {
-	return &osVersionListResultPager{
-		pipeline: client.con.Pipeline(),
-		requester: func(ctx context.Context) (*azcore.Request, error) {
+func (client *CloudServiceOperatingSystemsClient) ListOSVersions(location string, options *CloudServiceOperatingSystemsListOSVersionsOptions) *CloudServiceOperatingSystemsListOSVersionsPager {
+	return &CloudServiceOperatingSystemsListOSVersionsPager{
+		client: client,
+		requester: func(ctx context.Context) (*policy.Request, error) {
 			return client.listOSVersionsCreateRequest(ctx, location, options)
 		},
-		responder: client.listOSVersionsHandleResponse,
-		errorer:   client.listOSVersionsHandleError,
-		advancer: func(ctx context.Context, resp OSVersionListResultResponse) (*azcore.Request, error) {
-			return azcore.NewRequest(ctx, http.MethodGet, *resp.OSVersionListResult.NextLink)
+		advancer: func(ctx context.Context, resp CloudServiceOperatingSystemsListOSVersionsResponse) (*policy.Request, error) {
+			return runtime.NewRequest(ctx, http.MethodGet, *resp.OSVersionListResult.NextLink)
 		},
-		statusCodes: []int{http.StatusOK},
 	}
 }
 
 // listOSVersionsCreateRequest creates the ListOSVersions request.
-func (client *CloudServiceOperatingSystemsClient) listOSVersionsCreateRequest(ctx context.Context, location string, options *CloudServiceOperatingSystemsListOSVersionsOptions) (*azcore.Request, error) {
+func (client *CloudServiceOperatingSystemsClient) listOSVersionsCreateRequest(ctx context.Context, location string, options *CloudServiceOperatingSystemsListOSVersionsOptions) (*policy.Request, error) {
 	urlPath := "/subscriptions/{subscriptionId}/providers/Microsoft.Compute/locations/{location}/cloudServiceOsVersions"
 	if location == "" {
 		return nil, errors.New("parameter location cannot be empty")
@@ -256,36 +251,35 @@ func (client *CloudServiceOperatingSystemsClient) listOSVersionsCreateRequest(ct
 		return nil, errors.New("parameter client.subscriptionID cannot be empty")
 	}
 	urlPath = strings.ReplaceAll(urlPath, "{subscriptionId}", url.PathEscape(client.subscriptionID))
-	req, err := azcore.NewRequest(ctx, http.MethodGet, azcore.JoinPaths(client.con.Endpoint(), urlPath))
+	req, err := runtime.NewRequest(ctx, http.MethodGet, runtime.JoinPaths(client.ep, urlPath))
 	if err != nil {
 		return nil, err
 	}
-	req.Telemetry(telemetryInfo)
-	reqQP := req.URL.Query()
+	reqQP := req.Raw().URL.Query()
 	reqQP.Set("api-version", "2021-03-01")
-	req.URL.RawQuery = reqQP.Encode()
-	req.Header.Set("Accept", "application/json")
+	req.Raw().URL.RawQuery = reqQP.Encode()
+	req.Raw().Header.Set("Accept", "application/json")
 	return req, nil
 }
 
 // listOSVersionsHandleResponse handles the ListOSVersions response.
-func (client *CloudServiceOperatingSystemsClient) listOSVersionsHandleResponse(resp *azcore.Response) (OSVersionListResultResponse, error) {
-	var val *OSVersionListResult
-	if err := resp.UnmarshalAsJSON(&val); err != nil {
-		return OSVersionListResultResponse{}, err
+func (client *CloudServiceOperatingSystemsClient) listOSVersionsHandleResponse(resp *http.Response) (CloudServiceOperatingSystemsListOSVersionsResponse, error) {
+	result := CloudServiceOperatingSystemsListOSVersionsResponse{RawResponse: resp}
+	if err := runtime.UnmarshalAsJSON(resp, &result.OSVersionListResult); err != nil {
+		return CloudServiceOperatingSystemsListOSVersionsResponse{}, err
 	}
-	return OSVersionListResultResponse{RawResponse: resp.Response, OSVersionListResult: val}, nil
+	return result, nil
 }
 
 // listOSVersionsHandleError handles the ListOSVersions error response.
-func (client *CloudServiceOperatingSystemsClient) listOSVersionsHandleError(resp *azcore.Response) error {
-	body, err := resp.Payload()
+func (client *CloudServiceOperatingSystemsClient) listOSVersionsHandleError(resp *http.Response) error {
+	body, err := runtime.Payload(resp)
 	if err != nil {
-		return azcore.NewResponseError(err, resp.Response)
+		return runtime.NewResponseError(err, resp)
 	}
 	errType := CloudError{raw: string(body)}
-	if err := resp.UnmarshalAsJSON(&errType); err != nil {
-		return azcore.NewResponseError(fmt.Errorf("%s\n%s", string(body), err), resp.Response)
+	if err := runtime.UnmarshalAsJSON(resp, &errType); err != nil {
+		return runtime.NewResponseError(fmt.Errorf("%s\n%s", string(body), err), resp)
 	}
-	return azcore.NewResponseError(&errType, resp.Response)
+	return runtime.NewResponseError(&errType, resp)
 }
