@@ -7,10 +7,12 @@
 package recording
 
 import (
+	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"net/http"
 	"os"
+	"path/filepath"
 	"strings"
 	"testing"
 	"time"
@@ -368,7 +370,191 @@ func (s *recordingTests) TestRecordRequestsAndFailMatchingForMissingRecording() 
 }
 
 func (s *recordingTests) TearDownSuite() {
-	// cleanup test files
-	err := os.RemoveAll("recordings")
-	require.Nil(s.T(), err)
+	files, err := filepath.Glob("recordings/**/*.yaml")
+	require.NoError(s.T(), err)
+	for _, f := range files {
+		err := os.Remove(f)
+		require.NoError(s.T(), err)
+	}
+}
+
+func TestRecordingOptions(t *testing.T) {
+	r := RecordingOptions{
+		UseHTTPS: true,
+	}
+	require.Equal(t, r.HostScheme(), "https://localhost:5001")
+
+	r.UseHTTPS = false
+	require.Equal(t, r.HostScheme(), "http://localhost:5000")
+
+	require.Equal(t, GetEnvVariable(t, "Nonexistentevnvar", "somefakevalue"), "somefakevalue")
+	require.NotEqual(t, GetEnvVariable(t, "PROXY_CERT", "fake/path/to/proxycert"), "fake/path/to/proxycert")
+
+	r.Init()
+	require.Equal(t, r.Host, "localhost:5000")
+	require.Equal(t, r.Scheme, "http")
+
+	r.UseHTTPS = true
+	r.Init()
+	require.Equal(t, r.Host, "localhost:5001")
+	require.Equal(t, r.Scheme, "https")
+}
+
+var packagePath = "sdk/internal/recording"
+
+func TestStartStop(t *testing.T) {
+	os.Setenv("AZURE_RECORD_MODE", "record")
+	defer os.Unsetenv("AZURE_RECORD_MODE")
+
+	err := StartRecording(t, packagePath, nil)
+	require.NoError(t, err)
+
+	client, err := GetHTTPClient(t)
+	require.NoError(t, err)
+
+	req, err := http.NewRequest("POST", "https://localhost:5001", nil)
+	require.NoError(t, err)
+
+	req.Header.Set(UpstreamUriHeader, "https://www.bing.com/")
+	req.Header.Set(ModeHeader, GetRecordMode())
+	req.Header.Set(IdHeader, GetRecordingId(t))
+
+	resp, err := client.Do(req)
+	require.NoError(t, err)
+	require.NotNil(t, resp)
+
+	require.NotNil(t, GetRecordingId(t))
+
+	err = StopRecording(t, nil)
+	require.NoError(t, err)
+
+	// Make sure the file is there
+	jsonFile, err := os.Open("./recordings/TestStartStop.json")
+	require.NoError(t, err)
+	defer jsonFile.Close()
+}
+
+func TestUriSanitizer(t *testing.T) {
+	os.Setenv("AZURE_RECORD_MODE", "record")
+	defer os.Unsetenv("AZURE_RECORD_MODE")
+
+	err := StartRecording(t, packagePath, nil)
+	require.NoError(t, err)
+
+	err = AddUriSanitizer("replacement", "bing", nil)
+	require.NoError(t, err)
+
+	client, err := GetHTTPClient(t)
+	require.NoError(t, err)
+
+	req, err := http.NewRequest("POST", "https://localhost:5001", nil)
+	require.NoError(t, err)
+
+	req.Header.Set(UpstreamUriHeader, "https://www.bing.com/")
+	req.Header.Set(ModeHeader, GetRecordMode())
+	req.Header.Set(IdHeader, GetRecordingId(t))
+
+	resp, err := client.Do(req)
+	require.NoError(t, err)
+	require.NotNil(t, resp)
+
+	require.NotNil(t, GetRecordingId(t))
+
+	err = StopRecording(t, nil)
+	require.NoError(t, err)
+
+	// Make sure the file is there
+	jsonFile, err := os.Open("./recordings/TestUriSanitizer.json")
+	require.NoError(t, err)
+	defer jsonFile.Close()
+
+	var data RecordingFileStruct
+	byteValue, err := ioutil.ReadAll(jsonFile)
+	require.NoError(t, err)
+	err = json.Unmarshal(byteValue, &data)
+	require.NoError(t, err)
+
+	require.Equal(t, data.Entries[0].RequestUri, "https://www.replacement.com/")
+}
+
+func TestProxyCert(t *testing.T) {
+	_, err := getRootCas(t)
+	require.NoError(t, err)
+
+	tempProxyCert, ok := os.LookupEnv("PROXY_CERT")
+	require.True(t, ok)
+	err = os.Unsetenv("PROXY_CERT")
+	require.NoError(t, err)
+
+	_, err = getRootCas(t)
+	require.NoError(t, err)
+
+	err = os.Setenv("PROXY_CERT", "not/a/path.crt")
+	require.NoError(t, err)
+	_, err = GetHTTPClient(t)
+	require.Error(t, err)
+
+	os.Setenv("PROXY_CERT", tempProxyCert)
+}
+
+func TestStopRecordingNoStart(t *testing.T) {
+	os.Setenv("AZURE_RECORD_MODE", "record")
+	defer os.Unsetenv("AZURE_RECORD_MODE")
+
+	err := StopRecording(t, nil)
+	require.Error(t, err)
+
+	jsonFile, err := os.Open("./recordings/TestStopRecordingNoStart.json")
+	require.Error(t, err)
+	defer jsonFile.Close()
+}
+
+type RecordingFileStruct struct {
+	Entries []Entry `json:"Entries"`
+}
+
+type Entry struct {
+	RequestUri string `json:"RequestUri"`
+}
+
+func TestLiveModeOnly(t *testing.T) {
+	LiveOnly(t)
+	if GetRecordMode() == modePlayback {
+		t.Fatalf("Test should not run in playback")
+	}
+}
+
+func TestSleep(t *testing.T) {
+	start := time.Now()
+	Sleep(time.Second * 5)
+	duration := time.Since(start)
+	if GetRecordMode() == modePlayback {
+		if duration > (time.Second * 1) {
+			t.Fatalf("Sleep took longer than five seconds")
+		}
+	} else {
+		if duration < (time.Second * 1) {
+			t.Fatalf("Sleep took less than five seconds")
+		}
+	}
+}
+
+func TestBadAzureRecordMode(t *testing.T) {
+	temp := recordMode
+
+	recordMode = "badvalue"
+	err := StartRecording(t, packagePath, nil)
+	require.Error(t, err)
+
+	recordMode = temp
+}
+
+func TestBackwardSlashPath(t *testing.T) {
+	os.Setenv("AZURE_RECORD_MODE", "record")
+	defer os.Unsetenv("AZURE_RECORD_MODE")
+
+	packagePathBackslash := "sdk\\internal\\recordings"
+
+	err := StartRecording(t, packagePathBackslash, nil)
+	require.Error(t, err)
 }
